+++ src/main/java/NeuralNetwork/Layers/Layer.java
@@ -15,7 +15,7 @@ public class Layer implements ILayer {
 
     public void addNeuron(final Neuron neuron) {
         if (!this.neurons.isEmpty()) {
-            final long weightsSize = this.neurons.getLast().getWeightsSize();
+            final int weightsSize = this.neurons.getLast().getWeightsSize();
 
             if (neuron.getWeightsSize() != weightsSize) {
                 throw new IllegalArgumentException("New neuron weights size [" + neuron.getWeightsSize() + "] is not equal to current weights size [" + weightsSize + "]");
